requirements.txt
pandas
datetime
sqlalchemy
python-dotenv
paho-mqtt
requests
psycopg2